feat(auth): add initialization screen and update user preferences on login
This commit introduces an initialization screen to the application. This screen is responsible for determining whether the user is logged in or not. If the user is logged in, the user will be redirected to the home screen. Otherwise, the user will be redirected to the login screen.

Additionally, this commit updates the user's preferences in shared preferences upon successful login. This ensures that the user's preferences are always up to date.

These changes provide a more consistent and seamless user experience.