feat: Add loading state to primary button
This change introduces a loading state to the primary button, allowing it to display a circular progress indicator while the button is disabled.

The loading state is controlled by an `isLoading` RxBool, which can be set to `true` or `false` to indicate whether the button should be in a loading state. When the `isLoading` value is `true`, the button will be disabled and the progress indicator will be displayed. When the `isLoading` value is `false`, the button will be enabled and the progress indicator will be hidden.

This change improves the user experience by providing visual feedback to the user when the button is disabled due to a long-running operation.